+++ src/SmartBlaster.h
@@ -98,7 +98,7 @@ class SmartBlaster {
     //getters
     uint8_t getAmmo();
     uint32_t getChrono();
-    uint8_t getFlywheelPWM();
+    uint16_t getFlywheelPWM();
 
 
 };
